SDK-642: Move Response.cs to "Web" folder
