+++ test/Yoti.Auth.Tests/ProfileParserTests.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Yoti.Auth.Exceptions;
 using Yoti.Auth.Tests.TestTools;
+using Yoti.Auth.Web;
 
 namespace Yoti.Auth.Tests
 {
