+++ test/Yoti.Auth.Tests/YotiClientEngineTests.cs
@@ -9,6 +9,7 @@ using Yoti.Auth.Aml;
 using Yoti.Auth.Exceptions;
 using Yoti.Auth.ShareUrl;
 using Yoti.Auth.Tests.TestTools;
+using Yoti.Auth.Web;
 
 namespace Yoti.Auth.Tests
 {
